+++ CHANGELOG.md
@@ -5,6 +5,11 @@ New features:
 - Implement `ctrl-u` and `ctrl-d` commands to jump up and down by half
   the screen's height, or by a specified number of lines.
 
+Improvements:
+- Keep focused in same place on screen when toggling between line and
+  data modes; fix a crash when focused on a closing delimiter and
+  switching to data mode.
+
 v0.7.2 (2022-02-20)
 ==================